pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agents-sdk-pipeline"
version = "0.1.0"
description = "Graph pipeline for amendment analysis and DOCX redaction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.76.0",
    "python-docx>=1.1.2",
    "langgraph>=0.2.70",
    "langchain-core>=0.3.0",
]

[project.scripts]
graph-pipeline = "graph_pipeline.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
graph_pipeline = [
    "prompts/*.txt",
    "prompts/*.md",
    "amendment_analysis/prompts/*.txt",
]
